Hint order for FST files
